add predefined Luna requests for instart and ezadj
This commit adds predefined Luna requests that can be used to launch the
Instart and EZ Adjust menus. These requests provide an alternative way
to access the menus on older webOS versions where the existing keys
('i' and 'e', respectively) may not work.

The Luna requests should work on webOS 3 and newer. Note that on older
versions, the menus may be accessed through a different app
(com.webos.app.livetv).

The password for accessing the menus is likely to be 0413, but other
potential codes include 0000, 1105, and 7777.